+++ pkg/log/logger_manager.go
@@ -25,6 +25,7 @@ import (
 var (
 	DefaultLogger ErrorLogger
 	StartLogger   ErrorLogger
+	Proxy         ProxyLogger
 
 	ErrNoLoggerFound = errors.New("no logger found in logger manager")
 )
@@ -40,6 +41,8 @@ func init() {
 	StartLogger, _ = GetOrCreateDefaultErrorLogger("", INFO)
 	// default as start before Init
 	DefaultLogger = StartLogger
+	// default proxy logger for test, override after config parsed
+	Proxy, _ = CreateDefaultProxyLogger("", INFO)
 }
 
 // ErrorLoggerManager manages error log can be updated dynamicly
@@ -68,6 +71,14 @@ func (mng *ErrorLoggerManager) GetOrCreateErrorLogger(p string, level Level, f C
 	return lg, nil
 }
 
+func (mng *ErrorLoggerManager) SetAllErrorLoggerLevel(level Level) {
+	mng.mutex.Lock()
+	defer mng.mutex.Unlock()
+	for _, lg := range mng.managers {
+		lg.SetLogLevel(level)
+	}
+}
+
 // Default Export Functions
 func GetErrorLoggerManagerInstance() *ErrorLoggerManager {
 	return errorLoggerManagerInstance
@@ -80,6 +91,9 @@ func GetOrCreateDefaultErrorLogger(p string, level Level) (ErrorLogger, error) {
 
 func InitDefaultLogger(output string, level Level) (err error) {
 	DefaultLogger, err = GetOrCreateDefaultErrorLogger(output, level)
+	if err == nil {
+		Proxy, err = CreateDefaultProxyLogger(output, level)
+	}
 	return
 }