+++ pkg/protocol/rpc/engine.go
@@ -22,8 +22,6 @@ import (
 
 	"github.com/alipay/sofa-mosn/pkg/log"
 	"github.com/alipay/sofa-mosn/pkg/types"
-
-	mosnctx "github.com/alipay/sofa-mosn/pkg/context"
 )
 
 type engine struct {
@@ -67,7 +65,7 @@ func (eg *engine) BuildSpan(args ...interface{}) types.Span {
 	if builder != nil {
 		return builder.BuildSpan(args...)
 	} else {
-		log.DefaultLogger.Warnf("engine spanBuilder is empty,eg=%+v", eg)
+		log.DefaultLogger.Warnf("[protocol][engine][rpc] spanBuilder is empty, engine=%+v", eg)
 		return nil
 	}
 }
@@ -88,7 +86,7 @@ func (m *mixedEngine) Encode(ctx context.Context, model interface{}) (types.IoBu
 			return nil, ErrUnrecognizedCode
 		}
 	default:
-		log.DefaultLogger.Errorf("[protocol][engine] not RpcCmd, cannot find encoder for model = %+v", model)
+		log.DefaultLogger.Errorf("[protocol][engine][rpc] not RpcCmd, cannot find encoder for model = %+v", model)
 		return nil, ErrUnknownType
 	}
 }
@@ -99,7 +97,7 @@ func (m *mixedEngine) Decode(ctx context.Context, data types.IoBuffer) (interfac
 		code := data.Bytes()[0]
 
 		if log.DefaultLogger.GetLogLevel() >= log.DEBUG {
-			log.DefaultLogger.Debugf("mixed protocol engine decode, protocol code = %x", code)
+			log.DefaultLogger.Debugf("[protocol][engine][rpc] mixed engine decode, protocol code = %x", code)
 		}
 
 		if eg, exists := m.engineMap[code]; exists {
@@ -126,17 +124,16 @@ func (m *mixedEngine) Register(protocolCode byte, encoder types.Encoder, decoder
 }
 
 func (m *mixedEngine) BuildSpan(args ...interface{}) types.Span {
-	if len(args) == 0 {
+	if len(args) < 2 {
 		return nil
 	}
 
-	if _, ok := args[0].(context.Context); !ok {
+	cmd, ok := args[1].(RpcCmd)
+	if !ok || cmd == nil {
 		return nil
 	}
 
-	ctx := args[0].(context.Context)
-
-	engine := m.engineMap[mosnctx.Get(ctx, types.ContextSubProtocol).(byte)]
+	engine := m.engineMap[cmd.ProtocolCode()]
 
 	if engine == nil {
 		return nil