+++ pkg/proxy/downstream_test.go
@@ -73,8 +73,6 @@ func TestDownstream_FinishTracing_Enable_SpanIsNotNil(t *testing.T) {
 }
 
 func TestDirectResponse(t *testing.T) {
-	initGlobalStats()
-
 	testCases := []struct {
 		client *mockResponseSender
 		route  *mockRoute